+++ src/botvinnik/main.cpp
@@ -37,6 +37,7 @@
 #include "plugins/CheatSheet.hpp"
 #include "plugins/Debian.hpp"
 #include "plugins/Fortune.hpp"
+#include "plugins/Giphy.hpp"
 #include "plugins/LibreTranslate.hpp"
 #include "plugins/Ping.hpp"
 #include "plugins/Wikipedia.hpp"
@@ -254,6 +255,14 @@ int main(int argc, char** argv)
   {
     // Should never happen!
     std::cerr << "Error: Registration of translate plugin failed!" << std::endl
+              << "The bot will not start." << std::endl;
+    return bvn::rcPluginRegistrationError;
+  }
+  bvn::Giphy gif(config.gifApiKey(), bot.matrix());
+  if (!bot.registerPlugin(gif))
+  {
+    // Should never happen!
+    std::cerr << "Error: Registration of Giphy plugin failed!" << std::endl
               << "The bot will not start." << std::endl;
     return bvn::rcPluginRegistrationError;
   }