+++ tests/plugins/Giphy.cpp
@@ -1,0 +1,88 @@
+/*
+ -------------------------------------------------------------------------------
+    This file is part of the test suite for botvinnik.
+    Copyright (C) 2020, 2023  Dirk Stolle
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ -------------------------------------------------------------------------------
+*/
+
+#include "../locate_catch.hpp"
+#include "../../src/botvinnik/Bot.hpp"
+#include "../../src/conf/Configuration.hpp"
+#include "../../src/botvinnik/plugins/Giphy.hpp"
+
+TEST_CASE("plugin Giphy")
+{
+  using namespace bvn;
+  using namespace std::chrono;
+  Configuration conf;
+  Bot bot(conf);
+  Giphy plugin(conf.gifApiKey(), bot.matrix());
+
+  const auto commands = plugin.commands();
+
+  SECTION("commands must exist")
+  {
+    REQUIRE_FALSE( commands.empty() );
+
+    // giphy command has to be there.
+    REQUIRE_FALSE( std::find(commands.begin(), commands.end(), "giphy") == commands.end() );
+  }
+
+  SECTION("one line help")
+  {
+    for (const auto& cmd : commands)
+    {
+      // Help text must not be empty.
+      REQUIRE_FALSE( plugin.helpOneLine(cmd).empty() );
+    }
+  } // one line help section
+
+  SECTION("command handlers must return text")
+  {
+    const std::string_view mockUserId = "@alice:bob.charlie.tld";
+    const std::string_view mockRoomId = "!AbcDeFgHiJk345:bob.charlie.tld";
+    const milliseconds ts = duration_cast<milliseconds>(system_clock::now().time_since_epoch());
+    for (const auto& cmd : commands)
+    {
+      // Answer to commands must not be empty.
+      REQUIRE_FALSE( plugin.handleCommand(cmd, cmd, mockUserId, mockRoomId, ts).body.empty() );
+    }
+  }
+
+  SECTION("plugin registration")
+  {
+    // Plugin registration must be successful.
+    REQUIRE( bot.registerPlugin(plugin) );
+  }
+
+  SECTION("allowDeactivation()")
+  {
+    SECTION("non-core plugin commands can always be deactivated")
+    {
+      for (const auto& cmd : commands)
+      {
+        REQUIRE( plugin.allowDeactivation(cmd) );
+      }
+    }
+
+    SECTION("unknown commands / commands of other plugins cannot be deactivated")
+    {
+      REQUIRE_FALSE( plugin.allowDeactivation("foo") );
+      REQUIRE_FALSE( plugin.allowDeactivation("not-a-command") );
+      REQUIRE_FALSE( plugin.allowDeactivation("ping") );
+    }
+  }
+}