+++ tests/plugins/pluginRegistration.cpp
@@ -30,6 +30,7 @@
 #include "../../src/botvinnik/plugins/CheatSheet.hpp"
 #include "../../src/botvinnik/plugins/Debian.hpp"
 #include "../../src/botvinnik/plugins/Fortune.hpp"
+#include "../../src/botvinnik/plugins/Giphy.hpp"
 #include "../../src/botvinnik/plugins/LibreTranslate.hpp"
 #include "../../src/botvinnik/plugins/Ping.hpp"
 #include "../../src/botvinnik/plugins/Wikipedia.hpp"
@@ -51,6 +52,7 @@ TEST_CASE("plugin registration for all known plugins")
   Corona cov(bot.matrix());
   Debian deb;
   Fortune fortune;
+  Giphy giphy("key", bot.matrix());
   LibreTranslate translate("https://libretranslate.com", "");
   Ping ping(std::chrono::milliseconds(2345));
   Wikipedia wiki;
@@ -67,6 +69,7 @@ TEST_CASE("plugin registration for all known plugins")
     REQUIRE( bot.registerPlugin(cov) );
     REQUIRE( bot.registerPlugin(deb) );
     REQUIRE( bot.registerPlugin(fortune) );
+    REQUIRE( bot.registerPlugin(giphy) );
     REQUIRE( bot.registerPlugin(translate) );
     REQUIRE( bot.registerPlugin(ping) );
     REQUIRE( bot.registerPlugin(wiki) );